refactor(gui): extract log retention, redaction and export into logs.rs
Third and largest cut at lib.rs. Moves the eight log/file commands, the
DialogPathRegistry write gate, retention (prune_log_files / prune_trace_files),
the editor-opener detection, credential redaction, the zip export, and the two
test modules covering them (log_tools_tests, dialog_path_registry_tests).

These belong together because the AGENTS.md logging contract is stated across
them: batch-auth .trace files must never reach an export or archive zip, and the
export path redacts credentials while the archive path deliberately does not.
Those rules were previously spread over ~1400 lines of a 4000-line file with
their tests 1500 lines further down; the module header now states both
invariants next to the code and the tests that pin them.

Also fixes a doc comment I misplaced in the previous commit: sha256_hex had
landed between validate_archive_folder_name's doc and its function.

lib.rs 3935 -> 2522 (4630 at the start of this branch). 45 commands registered
and defined, unchanged; 362 workspace tests pass; clippy clean.
